fix: rename save filter to save rule
