+++ src/Domain/Item/SaveRule/ItemSaveRuleInterface.ts
