+++ src/Domain/Item/SaveRule/ItemSaveRuleResult.ts
